exception handling

+++ src/main/java/com/stackroute/muzixtrack/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.muzixtrack.service;
 
 import com.stackroute.muzixtrack.domain.Track;
+import com.stackroute.muzixtrack.exception.TrackAlreadyExistsException;
+import com.stackroute.muzixtrack.exception.TrackNotFoundException;
 import com.stackroute.muzixtrack.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,47 +20,76 @@ public class TrackServiceImpl implements TrackService {
   }
 
   @Override
-  public Track save(Track track) {
+  public Track save(Track track) throws TrackAlreadyExistsException {
+    if (trackRepository.existsById(track.getId())) {
+      throw new TrackAlreadyExistsException("track already exist");
+    }
     Track savedTrack = trackRepository.save(track);
+    if (savedTrack == null) {
+      throw new TrackAlreadyExistsException("track is null");
+    }
     return savedTrack;
   }
 
   @Override
-  public Track getById(int id) {
+  public Track getById(int id)throws TrackNotFoundException {
+    if (!trackRepository.existsById(id)) {
+      throw new TrackNotFoundException("track not found");
+    }
     Track retriveTrack = trackRepository.findById(id).get();
     return retriveTrack;
   }
 
 
   @Override
-  public List<Track> getAllTracks() {
-    return trackRepository.findAll();
+  public List<Track> getAllTracks()throws Exception {
+    if(trackRepository.findAll().isEmpty()){
+     throw new Exception("Internal server error");
+    }
+    else{
+      List<Track> allTracks=trackRepository.findAll();
+      return  allTracks;
+    }
+
 
   }
 
   @Override
-  public Optional<Track> deleteTrackById(int id) {
-    Optional<Track> optional = trackRepository.findById(id);
-    trackRepository.deleteById(id);
-    if (optional.isPresent()) {
+  public Track deleteTrackById(int id)throws TrackNotFoundException {
+    if(trackRepository.existsById(id))
+    {
+      Track retrivedTrack=trackRepository.findById(id).get();
+      trackRepository.deleteById(id);
+      return retrivedTrack;
     }
-    return optional;
+    else
+    {
+      throw new TrackNotFoundException("Track not found");
+    }
   }
 
   @Override
   //update a particular  part of a track by its id
-  public Track updateTrack(int id, Track track) {
+  public Track updateTrack(int id)throws TrackNotFoundException {
 //        delete the track
-    trackRepository.deleteById(id);
-//        edit the track and save it
-    Track updateTrackById = trackRepository.save(track);
-    return updateTrackById;
+    if(trackRepository.existsById(id)){
+      Track updateTrackById=trackRepository.findById(id).get();
+      return updateTrackById;
+    }
+    else{
+      throw new TrackNotFoundException("track not found");
+    }
   }
 
   //get the track by name
   @Override
-  public List<Track> trackByName(String name) {
-    return trackRepository.trackByName(name);
+  public List<Track> trackByName(String name)throws Exception {
+    if (trackRepository.trackByName(name).isEmpty()) {
+      throw new Exception("no track with this name");
+    } else {
+      List<Track> allTracksByName=trackRepository.trackByName(name);
+      return  allTracksByName;
+    }
   }
 
 }
